// src/types/practice.ts
export interface PracticeInfo {
  practiceName: string;
  practicePhone: string;
  practiceFax: string;
  facilityName: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zip: string;
  country: string;
  timeZone: string;
  observesDaylightSaving: boolean;
}